App.js calls the Locations (mocking tho)
This was a pain, doing this caused a thing that reloaded the api call every single time you did anything, mainly because its just how react works. I fought this for hours and hours until I did the simplest thing ever and fixed it. All together its actually better code now but yeah I hate it and it added a lot. Also we have a loading page now. Its super boring but shouldnt bee seen too much, I think I will make something more interesting to make it not look so ugly, but that is a stretch goal